refactor(node): extract follower sync module
