+++ crates/node/src/lib.rs
@@ -10,6 +10,7 @@ use eyre as _;
 pub mod cli;
 pub mod dev;
 pub mod engine;
+mod follower;
 pub mod genesis;
 pub mod node;
 mod replication;
